Move columns, filerts, searchKeys to cookieListing from tableContainer
